Update RSSParser.php
This is a new configuration that  is responsible for parsing the unsecure urls "http" to "https". I faced a problem,because my site is https and when I have "http" src of image from the rss, that was causing security warnings and my site was not completely secure. After these changes my site is secure again and user will not complain.